Rework silx.app.view.__init__ to avoid to load GUI library
